Rename Typescript implementations of script editor/form components to replace old JS versions
