+++ src/components/editors/ActorEditor.tsx
@@ -31,8 +31,8 @@ import DirectionPicker from "../forms/DirectionPicker";
 import { DMG_PALETTE } from "../../consts";
 import { SpriteSheetSelectButton } from "../forms/SpriteSheetSelectButton";
 import { WorldEditor } from "./WorldEditor";
-import ScriptEditorDropdownButton from "../script/ScriptEditorDropdownButton2";
-import ScriptEditor from "../script/ScriptEditor2";
+import ScriptEditorDropdownButton from "../script/ScriptEditorDropdownButton";
+import ScriptEditor from "../script/ScriptEditor";
 import { AnimationSpeedSelect } from "../forms/AnimationSpeedSelect";
 import { MovementSpeedSelect } from "../forms/MovementSpeedSelect";
 import CollisionMaskPicker from "../forms/CollisionMaskPicker";
