+++ src/components/script/ScriptEventFormInput.tsx
@@ -1,24 +1,24 @@
-import { ActorSelect } from "components/forms/ActorSelect2";
+import { ActorSelect } from "components/forms/ActorSelect";
 import { AnimationSpeedSelect } from "components/forms/AnimationSpeedSelect";
 import AnimationStateSelect from "components/forms/AnimationStateSelect";
 import { AvatarSelect } from "components/forms/AvatarSelect";
 import { BackgroundSelect } from "components/forms/BackgroundSelect";
-import { CameraSpeedSelect } from "components/forms/CameraSpeedSelect2";
+import { CameraSpeedSelect } from "components/forms/CameraSpeedSelect";
 import CollisionMaskPicker from "components/forms/CollisionMaskPicker";
 import { CustomEventSelect } from "components/forms/CustomEventSelect";
 import DirectionPicker from "components/forms/DirectionPicker";
 import { EmoteSelect } from "components/forms/EmoteSelect";
 import EngineFieldSelect from "components/forms/EngineFieldSelect";
-import { FadeSpeedSelect } from "components/forms/FadeSpeedSelect2";
+import { FadeSpeedSelect } from "components/forms/FadeSpeedSelect";
 import InputPicker from "components/forms/InputPicker";
 import { MovementSpeedSelect } from "components/forms/MovementSpeedSelect";
-import { MusicSelect } from "components/forms/MusicSelect2";
-import { OperatorSelect } from "components/forms/OperatorSelect2";
-import { OverlayColorSelect } from "components/forms/OverlayColorSelect2";
+import { MusicSelect } from "components/forms/MusicSelect";
+import { OperatorSelect } from "components/forms/OperatorSelect";
+import { OverlayColorSelect } from "components/forms/OverlayColorSelect";
 import { PaletteSelect } from "components/forms/PaletteSelect";
-import { PropertySelect } from "components/forms/PropertySelect2";
+import { PropertySelect } from "components/forms/PropertySelect";
 import { SceneSelect } from "components/forms/SceneSelect";
-import { SoundEffectSelect } from "components/forms/SoundEffectSelect2";
+import { SoundEffectSelect } from "components/forms/SoundEffectSelect";
 import { SpriteSheetSelect } from "components/forms/SpriteSheetSelect";
 import { VariableSelect } from "components/forms/VariableSelect";
 import castEventValue from "lib/helpers/castEventValue";
