+++ src/components/world/Connections.tsx
