+++ src/components/world/SceneInfo.tsx
